requirements.txt
beautifulsoup4==4.3.2
bleach==1.4.1
cffi==0.9.2
coverage==3.7.1
cryptography==0.8
decorator==4.0.4
Django==1.8.4
django-allauth==0.21.0
django-appconf==1.0.1
django-braces==1.4.0
django-date-extensions==0.1.dev0
django-debug-toolbar==1.3.2
django-debug-toolbar-template-timings==0.6.4
django-extensions==1.6.1
django-filter==0.11.0
django-haystack==2.4.1
django-model-utils==2.3.1
django-nose==1.4.1
django-pipeline==1.5.1
-e git+https://github.com/mysociety/django-popolo@update-with-better-migrations-redux#egg=mysociety-django-popolo
django-statici18n==1.1.5
django-webtest==1.7.7
djangorestframework==3.3.1
docutils==0.10
enum34==1.0.4
factory-boy==2.6.0
fake-factory==0.5.3
futures==3.0.3
html5lib==0.999
ipdb==0.8
ipython==4.0.0
ipython-genutils==0.1.0
jsonpatch==1.9
jsonpointer==1.6
lxml==3.4.1
mock==1.0.1
mysociety-django-images==0.0.5
ndg-httpsclient==0.3.3
nose==1.3.7
numpy==1.9.1
oauthlib==0.7.2
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
Pillow==2.7.0
psycopg2==2.5.4
pyasn1==0.1.7
pycparser==2.10
pyelasticsearch==1.4
pygeocoder==1.2.5
pyOpenSSL==0.14
python-dateutil==2.4.2
python-magic==0.4.6
python-memcached==1.54
python-openid==2.2.5
python-slugify==0.0.7
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
simplegeneric==0.8.1
six==1.7.3
sorl-thumbnail==12.3
sqlparse==0.1.15
traitlets==4.0.0
Unidecode==0.4.17
waitress==0.8.9
WebOb==1.4
WebTest==2.0.15
yanc==0.2.4
